Extracted form modal functionality to separate modals
