+++ frontend/src/styles/buttons.css
@@ -25,3 +25,12 @@ button {
     display: block;
     margin: 0 auto;
 }
+
+.form-button-spacing {
+    margin-top: 50px;
+    margin-bottom: 10px;
+}
+
+.form-button-size {
+    height: 50px;
+}
